#9: Fix return type for get entity set
